Reduced voxel size from 4 bytes to 1

+++ src/Engine.hpp
@@ -29,7 +29,7 @@ struct FrameData {
 };
 
 struct Voxel {
-    int colourIndex;
+    uint8_t colourIndex;
 };
 
 struct VoxelPushConstants {
